refactored code to class based
