+++ tap_google_sheets/discover.py
@@ -1,26 +1,40 @@
 from singer.catalog import Catalog, CatalogEntry, Schema
-from tap_google_sheets.schema import get_schemas, STREAMS
+from tap_google_sheets.schema import STREAMS
 
 
 def discover(client, spreadsheet_id):
-    schemas, field_metadata = get_schemas(client, spreadsheet_id)
     catalog = Catalog([])
 
-    for stream_name, schema_dict in schemas.items():
-        schema = Schema.from_dict(schema_dict)
-        mdata = field_metadata[stream_name]
-        key_properties = None
-        for mdt in mdata:
-            table_key_properties = mdt.get('metadata', {}).get('table-key-properties')
-            if table_key_properties:
-                key_properties = table_key_properties
+    for stream, stream_obj in STREAMS.items():
+        stream_object = stream_obj(client, spreadsheet_id)
+        schemas, field_metadata = stream_object.get_schemas()
 
-        catalog.streams.append(CatalogEntry(
-            stream=stream_name,
-            tap_stream_id=stream_name,
-            key_properties=STREAMS.get(stream_name, {}).get('key_properties', key_properties),
-            schema=schema,
-            metadata=mdata
-        ))
+        # loop over the schema and prepare catalog
+        for stream_name, schema_dict in schemas.items():
+
+            schema = Schema.from_dict(schema_dict)
+            mdata = field_metadata[stream_name]
+
+            key_properties = None
+
+            # get primary key for the stream
+            for mdt in mdata:
+                table_key_properties = mdt.get('metadata', {}).get('table-key-properties')
+                if table_key_properties:
+                    key_properties = table_key_properties
+
+            if not STREAMS.get(stream_name):
+                key_props = key_properties
+            else:
+                stream_obj = STREAMS.get(stream_name)(client, spreadsheet_id)
+                key_props = stream_obj.key_properties
+
+            catalog.streams.append(CatalogEntry(
+                stream=stream_name,
+                tap_stream_id=stream_name,
+                key_properties=key_props,
+                schema=schema,
+                metadata=mdata
+            ))
 
     return catalog
